XG_MaterialNode: bind() slot to size_t from uint32_t
+ Fix behavior of slot
+ Fix material struct buffer write

+++ GMFileViewer/FileTypes/XGM/XG/SubNodes/XG_MaterialNode.h
@@ -2,6 +2,6 @@
 class XG_MaterialNode
 {
 public:
-	virtual void bind(uint32_t slot) const = 0;
+	virtual void bind(size_t slot) const = 0;
 	virtual size_t getNumMaterials() const = 0;
 };
